Add empty declarations and checks to partly fix START_WITH_DAY.
This commit adds some variable declarations and other minor things so
that setting var.START_WITH_DAY to 1 doesn't cause errors. This commit
doesn't actually fix that gamemode completely, since I'm still
pondering how best to make the game notify people of their roles when
the game starts in this mode, but it's a good start.