Change playercounts for drunk, traitor, and wolf2
Per discussion in #wolfgame-dev, this commit changes the playercounts at
which drunk, traitor, and wolf2 are introduced. It removes drunk from 6p
and 7p games (moving it from 6 to 8), adds traitor to 8p and 9p games
(moving it from 10 to 8), and removes wolf2 from 8p and 9p games (moving
it from 8 to 10).

Or, in other words, it moves drunk later and switches traitor and wolf2.

Statistics on the effect of these two changes can be measured
independently, because they affect different ranges of games.

If this commit causes gameplay or other issues, feel free to revert it.