Move various in-channel messages to user notices
This commit moves the !votes and !stats rate limit warnings to user
notices intead of in-channel messages, to be consistent with the other
rate limit warnings and because those messages are usually just
pointless noise. It also moves the CARE_BOLD and CARE_COLOR messages
to user notices, because they're spammy and annoying and just make
things worse when they're in-channel.