Change branch condition for checking if night is over, fixes #17
chk_nightdone used to check whether the number of people seen equals
the number of seers (and the same for harlot, etc.). This means night
times out if the seer (etc.) leaves after seeing but before night is
otherwise ended. This commit changes equals to >=, and swaps around the
wolf part of that condition so that it matches the other parts.