Fix werecrow dying when wolf is shot dead at night
Previously, if a wolf was shot dead at night, any werecrows that have
observed were removed from the game without announcement, although they
would remain voiced.

For those who are curious, this is because lists are copied by reference
on assignment in Python; any changes made on the variable you assigned
the list to will be reflected in the original list. This can be avoided
by making a shallow copy of the list using `list(x)` or `x[:]`.