Lots of end-game message stuff.
The lover of a fool can win with the fool. The lover of a monster can win with monsters.
List winners at end of game.
Idlers are the scum of the earth and never win.
Displays a message if everyone died (originally by nyuszika7h, replicated here due to merge conflicts).