Fix issue where bot would spam services
This very simple fix probably seems redundant or even entirely
pointless, but it actually solves a big design flaw. It does not show
much for freenode, but for other networks (e.g. Esper), where I tweaked
my bot to use ns_ghost instead, it keeps trying to kill off an
(unregistered) nick, without having a password. This effectively causes
spam to the services, and results in an automatic ban.