Remove jester from default role set
The inclusion of jester was causing a disproportionately high number of wolf wins due to wolves voting jester as a strategy in order to bring numbers to even. Per discussion in -ops, it was decided that removing jester would be good in order to prevent this (an alternative option being reducing number of wolves so that the gameplay can still happen, but it requires more finesse on the wolves' part in order to utilize).