Always initialize opt-in ping list
If OPT_IN_PING is enabled via botconfig.py, OPT_IN_PING isn't set to
True at the time we set up tables as necessary, so people using !in will
make the bot run into the lack of table and crash.

We already do this with the away list even if we're going to use opt-in
ping, so this shouldn't cause any issues.