Announce wolf actions in wolfchat
This way it's immediately obvious if it was a legit command or not. The
result of werecrow's observation is not announced, because it's possible
for the crow to be killed at night and this would take away that
strategy. For sorcerer, it does not even bother announcing the fact that
they observed, since they can tell the result immediately if they are
alive, so the message is pretty much redundant.

Also, minor fixes:

* Log kill by angry wolves properly, instead of the non-obvious hack
  that @Vgr255 used.

* Add missing quotes in angry wolves wolfchat message.