Group all night messages together
This groups messages together for purposes of role + assassin and some
potential future template. I have purposefully left minion out, as the
role does not quite fit the same rules as the others. It will send two
different lists if they are also assassin, but we want that, NOT the
opposite! Doctor also now gets a player list, which may very well be
outdated by daybreak, but detective gets player list too so why not
doctor.