also fix 13p aleatoire
I totally misread stats and thought that village needed a boost at 14, whereas the boost was actually needed at 13. At the same time, 14 was relatively good beforehand, so needed to reintroduce something there that advantaged wolves. This may turn out to be too large of a boost, but we shall see.