Show local variables from all possibly relevant frames
This doesn't show the variables for a few specific functions, specifically those which have a '_ignore_locals_' local variable set to a truthy value. For now, those functions are handle_error.__call__ (which is present in every traceback ever, due to its nature), cmd.caller and command.caller (one or the other is present in a large number of tracebacks, again due to their nature).

This makes debugging easier when the error is not in the innermost frame, but rather in one or two frames before, and the locals are lost.