Add the new User Containers
This adds the new User Containers and the first steps of proper handling for it. Eventually, all of the containers dealing with users will be converted over to this. The ultimate goal is to be able to ask users to replace themselves with another user, and for it to be done transparently.