Add centralized error reporting (off by default)
If enabled, sends error details over to us so that we can look at them.
Includes a lot more details than the standard pastebin does, including
full IRC logs for the game as well as various game state vars that lets
us see hidden stuff more easily.

Still a work in progress, could probably be expanded to be more useful.
Compressing the data we stream over is probably not a bad idea either...