Revamp ROLE_GUIDE
ROLE_GUIDE is no longer a matrix and the list of valid roles is no
longer pulled from it. This allows for a large amount of cleanup.

This commit is focused on updating the usages in gamemodes.py; code
which makes use of ROLE_GUIDE elsewhere will be updated in a future
commit.