Update wild child to use the new_role and swap_role_state events
This isn't tested, and isn't (necessarily) expected to work. This is the role that prevented exchange from working, and this is what I think is the proper fix. Eventually, I'll rework the exchange machinery and test this change then. Until then, this is pretty much dead code and doesn't matter.

The wild-child-wolf would get "You are now a wild child" followed immediately by "Your idol has died, you are now a wolf". It's a minor annoyance, but I don't think it really matters much.