Allow multiple thresholds for !pingif
Removed var.PINGED_ALREADY and var.PINGED_ALREADY_ACCS as that wouldn't work with multiple thresholds. Instead replaced it by var.LAST_PING which keeps track of when the last ping was sent.

This required a DB schema upgrade, to coerce the pingif field from INTEGER to TEXT. It is now a comma-separated list of integers.