Make sure to apply misdirection sequentially
Previously I explicitly did it sequentially, but misdirection as actor applies first, and the original target may be different by the time you reach misdirection as target, which is why it's important to do it sequentially. In that case, it may result in the same target as initially, or two targets over in either direction. Either way, this is the proper way to do it.