Fix IDE compatibility for cats and statuses
By dynamically munging __all__, certain IDEs (such as PyCharm) do not
pick up on the changes we've made to it and instead treat our code as if
it has errors due to symbols not being defined. Make things more
explicit instead of being clever, even though it does mean more code in
the files and a bit of extra boilerplate down the line if adding
additional cats or statuses. The tradeoff is worthwhile.