Bot ident fixes
The "loggedin" message (and possibly others) may report back that we
have an ident without a leading ~, even if we do not have identd
running. Later on, it will include the leading ~ in our ident, however
the bot was convinced that we didn't have that so things were breaking
(namely, we never executed /WHO on joined channels so the bot was
unresponsive to all commands as the user list was never initialized).

This behavior was observed on freenode, but implemented as a generic
workaround in case other ircds misbehave in the same fashion.