Expand workaround to fix up our ident
When the ircd sends event_hosthidden, that only includes host but not
ident info, meaning that we're only partially set up in the bot.

We need to be able to fix up our ident in all cases (such as when it's
missing entirely), so expanding the ~ workaround to include that seems
like a sane idea.