Lover balance fix
In order to win as lover, you must both stay alive until end of game
*and* one of the lovers must have received a team win.

The second condition used to not exist, and this lead to a lot of abuse
by two vil-lovers joining wolfteam and derailing the game day 1.

Lover win condition is in addition to regular win conditions, so you can
still win with your team as per normal (regardless of whether you are
alive at the end of game or not).