Make var.DISCONNECTED a UserDict
This is meant to store people who have disconnected from the game for
any reason (part, quit, account change, etc.). In order for them to
return to the village properly, however, we need to keep track of
changes we know were made to them while they were "disconnected".

The User.disconnected property is meant to specify whether they're fully
off of IRC or not, but it seems to not be properly used inside of lykos
which is causing other bugs. That has not been fixed as part of this
commit.