Some backend cleanup of things
Make our featurelist handling more robust by normalizing values as the
bot needs them (since features can be delivered in arbitrary order, we
need to store what the server actually sent). The end goal here is to
move away from using dict keys and into using attributes directly for
enhanced type checking capability.

Recognize CPRIVMSG and CNOTICE. A new configuration DISABLE_CPRIVMSG is
provided so bot operators can disable this handling in the event it is
broken or if it's implicitly part of PRIVMSG/NOTICE already (as is the
case on freenode).

Add a new event irc_connected that bot operators can listen to in custom
code to perform whatever stuff they need as part of the connection
process. Fixes #435.