Finish up botconfig conversion (mostly)
Still need to un-tangle the mess that is src.settings/var and handle
things a bit more sanely across the board. For now we'll still assume
there's only one IRC transport and it's at index 0; that will change in
the future as we support more stuff.

Quite a long todo list of things before this can really be merged back
into master. Help wanted, apply in #lykos :)