Restrict wolfchat access to main roles
If someone is a wolf role as a secondary role but not main role, then
they do not win with wolves but have access to whatever wolf commands
their secondary role grants. As such, don't give them access to wolfchat
or relayed commands from other wolves.

Wolflist was already restricted to main roles only and didn't need to be
changed.