Convert DAY_COUNT and NIGHT_COUNT
Also modify the way night and day are counted, now incrementing only at the end of the transition_* events, before the begin_* event. Most roles were checking for NIGHT_COUNT == 0 (or similar), which translates to "we are not transitioning into night". This was non-obvious at first glance, and the new code makes that much cleaner and more obvious.