Kaboom balance fix
Wolf needed a lot of help in this mode, so let's make one of them
blessed. At later counts, we ensure that the wolf oracle is *not* also
the blessed wolf.

If this swings too far in wolves' favor, next up is probably to make
master of teleportation a village-aligned role instead of neutral. I
might end up doing that anyway just so there aren't too many roles with
the "must survive to win" condition in a mode with a lot of splash
damage from MS and VG.