Allow fstasis to add stasis
This reverts a previous decision I arbitrarily made to require that any
additions to a person's stasis count go through the !fwarn command.
There may be times where it is valuable or necessary to increase stasis
without issuing a full-on warning (perhaps because someone only has the
+A flag and therefore can't issue warnings, or because it needs to be
done very quickly and the !fwarn command is approximately as easy to use
and remember as tar).