Migrate module state to GameState (POC)
This involves some crimes with dynamically creating a type that extends
all known game state extensions (such as that found in matchmaker, which
was converted as a proof of concept). Some boilerplate is needed so that
static type analysis properly registers the combined game state class,
at least insofar as they need to care. This was tested on PyCharm, VS
Code, and MyPy and works on all 3.